setup_raw_data.py: do not fail if no udm files are available
the central function to gather the images is get_mask_images()
this function raises a ValueError if it does not find udm-files,
although they are not necessary most of the time. optionally
supress the Error with a parameter.

Also just don't mask the images if no udm file is found.